+++ projects/adc_dac/main.yml
@@ -1,13 +1,13 @@
 ---
-project: blink
+project: adc_dac
 parent: default
 board: red-pitaya
 
 cores:
-  - redp_adc_v1_0
-  - redp_dac_v1_0
   - axi_cfg_register_v1_0
   - axi_sts_register_v1_0
+  - redp_adc_v1_0
+  - redp_dac_v1_0
 
 memory:
   - name: config
